backend/requirements.txt
flask>=3.0
flask-cors>=4.0
torch>=2.2
torchvision>=0.17
timm>=1.0
Pillow>=10.0
